+++ app/data/events/index.js
@@ -1,6 +1,6 @@
 const { getEventsByFrn } = require('./frn/get-events-by-frn')
 const { getEventsByCorrelationId } = require('./correlation-id/get-events-by-correlation-id')
-const { getEventsBySchemeId } = require('./scheme-id/get-events-by-scheme')
+const { getEventsBySchemeId } = require('./scheme-id/get-events-by-scheme-id')
 
 module.exports = {
   getEventsByFrn,
